Moves JointFitExperiments to collections
